+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 
 * feat: add static code diagnostic [`avoid-non-ascii-symbols`](https://dartcodemetrics.dev/docs/rules/common/avoid-non-ascii-symbols).
+* feat: remove declaration in [`prefer-immediate-return`](https://dartcodemetrics.dev/docs/rules/common/prefer-immediate-return).
+* fix: correctly handle disabling rules with false.
 * fix: dart-code-metrics crash saying `Bad state: No element` when running command.
 
 ## 4.14.0